server/requirements-minimal.txt
# Minimal requirements for testing - use this if the main requirements.txt fails
# Copy this to requirements.txt if needed

setuptools==69.5.1
wheel==0.43.0
pip==24.0

Flask==2.3.3
Flask-Cors==4.0.0
gunicorn==21.2.0

requests==2.31.0
urllib3==1.26.18
certifi==2023.7.22

numpy==1.24.3
Pillow==10.0.1

# Note: opencv-python-headless and inference are commented out for testing
# Uncomment one by one to identify the problematic package

# opencv-python-headless==4.8.1.78
# inference==0.9.13
